refactored app into smaller components, changed file structure, removed template boilerplate
