+++ src/App.js
@@ -1,17 +1,12 @@
 import React from 'react'
-import { Counter } from './features/counter/Counter'
-import './App.css'
-import SignUp from './features/SignUpForm/SignUpOverlay';
-import {BrowserRouter as Router, Route} from "react-router-dom"
-
+import SignUp from './features/signUpForm/SignUpOverlay'
+import { Route } from 'react-router-dom'
 
 function App() {
   return (
-    <div>
-      <header>
-        <Route exact path ="/signup" component={SignUp}/>
-      </header>
-    </div>
+    <>
+      <Route exact path='/signup' component={SignUp} />
+    </>
   )
 }
