+++ src/App.test.js
@@ -1,15 +1,13 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { Provider } from 'react-redux';
-import store from './store';
-import App from './App';
+import React from 'react'
+import { render } from '@testing-library/react'
+import { Provider } from 'react-redux'
+import store from './redux/store'
+import App from './App'
 
-test('renders learn react link', () => {
-  const { getByText } = render(
+test('renders App without crashing', () => {
+  render(
     <Provider store={store}>
       <App />
     </Provider>
-  );
-
-  expect(getByText(/learn/i)).toBeInTheDocument();
-});
+  )
+})
