--- src/features/SignUpForm/ReviewInfo.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-
-
-export default function Review() {
-
-    return(
-        null
-    )
-
-}
