--- src/store.js
@@ -1,8 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import counterReducer from './features/counter/counterSlice';
-
-export default configureStore({
-  reducer: {
-    counter: counterReducer,
-  },
-});
